+++ src/app/locales/en/components/providers-list.js
@@ -1,0 +1,7 @@
+export default {
+  providers: 'Oneproviders',
+  providerItem: {
+    supportedSpaces: 'Number of supported spaces: {{count}}',
+    spaceSupportSize: 'Support provided: {{size}}',
+  },
+};