+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-6803 Added Oneprovider info popup
 * VFS-8640 Updated commons
 * VFS-8874 Removed bower and ember-spin-button
 * VFS-6397 Removed redundant bower dependencies